build(mypy): the gate enforces "no implicit Any" instead of asserting it
CLAUDE.md bans implicit Any; the config enabled two flags and a global
ignore_missing_imports, so `mypy .` was green while --strict reported 123
errors. Worse, the blanket flag was also swallowing the test tree's own
helpers/e2e/unit imports — every shared oracle helper resolved to Any at
every call site, unchecked.

The gate is now strict = true with mypy_path = src:tests, and the two
genuine exemptions are named rather than global: the six external packages
that ship no stubs, and the GNU Radio subclassing seam (gr.basic_block is
Any, so blocks extending it cannot satisfy disallow_subclassing_any).

Fixing the 123: CompileError imports point at compile/errors.py instead of
laundering through the compiler (24 sites, one fake home gone); 4 dead
type: ignore comments deleted and 3 live ones replaced by the isinstance
narrowing that made them unnecessary — 5 remain, 1 justified inline;
tests/helpers/parse.py grows ParseCase + a Message alias so case specs are
validated instead of splatted; CrcSpec owns the CRC parameters that were
riding a **dict into bool/str keywords; numpy expressions bind precisely
typed locals rather than widening to Any at 30 return sites.

1382 pass, 9 skip, 1 xfail. The only remaining suppression states why.

+++ src/marconi/engine/backends/gnuradio/embedded/chirp.py
@@ -151,7 +151,7 @@ def _parabolic(f: npt.NDArray[np.floating[Any]], p: int) -> float:
     if 0 < p < len(f) - 1:
         denom = f[p - 1] - 2.0 * f[p] + f[p + 1]
         if denom != 0.0:
-            return p + 0.5 * (f[p - 1] - f[p + 1]) / denom
+            return float(p + 0.5 * (f[p - 1] - f[p + 1]) / denom)
     return float(p)
